requirements.txt
streamlit
jinja2
pandas
python-dotenv
requests
openai
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
google-cloud-bigquery
hashlib
phonenumbers